+++ test/agenda/cancel.js
@@ -1,5 +1,5 @@
 import test from 'ava';
-import {beforeEach, afterEach} from '../helpers';
+import {beforeEach, afterEach} from '../helpers/agenda';
 
 test.beforeEach(beforeEach);
 test.afterEach.always(afterEach);
