+++ test/agenda/unique.js
@@ -1,7 +1,7 @@
 import test from 'ava';
 import delay from 'delay';
 import {Job} from '../../lib';
-import {beforeEach, afterEach} from '../helpers';
+import {beforeEach, afterEach} from '../helpers/agenda';
 
 test.beforeEach(beforeEach);
 test.afterEach.always(afterEach);
